Rename the file to English and delete the previous employee's files.
