+++ feature/plan/Contextual_Search_Main.py
@@ -1,11 +1,11 @@
-from feature.plan.utils.篩選條件.check_class import filter_by_label_type
-from feature.plan.utils.篩選條件.檢查預算 import filter_by_budget
-from feature.plan.utils.篩選條件.檢查距離 import filter_by_distance
-from feature.plan.utils.篩選條件.檢查日期 import filter_by_weekday
-from feature.plan.utils.篩選條件.檢查時間1 import filter_by_time_without_weekday
-from feature.plan.utils.標準化.相似性標準化point import normalize_similarity
-from feature.plan.utils.標準化.總評論標準化point import normalize_and_match, load_extracted_data
-from feature.plan.utils.標準化.距離的反向標準化point import calculate_reverse_normalized_distances_no_threshold
+from feature.plan.utils.Filter_Criteria.check_class import filter_by_label_type
+from feature.plan.utils.Filter_Criteria.check_budget import filter_by_budget
+from feature.plan.utils.Filter_Criteria.check_distance import filter_by_distance
+from feature.plan.utils.Filter_Criteria.check_date import filter_by_weekday
+from feature.plan.utils.Filter_Criteria.check_time import filter_by_time_without_weekday
+from feature.plan.utils.Norma_lization.similarity_score_normalized import normalize_similarity
+from feature.plan.utils.Norma_lization.comment_score_normalized import normalize_and_match, load_extracted_data
+from feature.plan.utils.Norma_lization.distances_score_normalized import calculate_reverse_normalized_distances_no_threshold
 
 
 def main(points, user_requirements):
