--- feature/plan/test.py
@@ -1,5 +1,0 @@
-
-
-import feature
-
-print(feature.__file__)
