+++ feature/plan/utils/fake_impormation
